requirements.txt
astroid==2.0.2
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
pkg-resources==0.0.0
pylint==2.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
typed-ast==1.1.0
typing==3.6.4
urllib3==1.23
wrapt==1.10.11
